requirements.txt
beautifulsoup4
streamlit
requests
shelve
python-dotenv
langchain
langchain-openai
langchain-core
langchain-text-splitters